Reimplement dynamic metadata handling
This has been tested for both playback of streams and
outputting to streams, and seems to work fine with minimal
locking.  This reuses the sequence number infrastructure
in OutputBuffer for synchronizing metadata payloads; so
(IMNSHO) should be much more understandable than various
flags being set here and there..

It could still use some cleanup and much testing, but
synchronization issues should be minimal.