playlist: re-randomize when explicitly playing a new song
When random is enabled and a user explicitly specifies
a certain song on the playlist should be played; we need
to re-randomize the internal ordering.

To reproduce this, assuming a four song playlist:

  play <song_a>
  next => <song_b>
  next => <song_c>
  next => <song_d>
  play <song_a>
  next => <song_b>
  next => <song_c>
  next => <song_d>
  ...

That is, the "next" command restarts song_{b,c,d} the second
time "play" starts playing song_a.

Thus, the second time "play" is called, the ordering of
song_{b,c,d} needs to be reshuffled.

Reported-by: Qball