alsa: use blocking instead of non-blocking write
The way we used non-blocking mode was HORRIBLE.

It was non-blocking to ALSA, but we end up blocking in a busy
loop that does absolutely NOTHING but retry.  We don't check
for playback cancellation (like we do in decoders) or anything.

This is seriously broken and I can imagine it affects people on
fast CPUs more because we do asynchronous output buffering and
our ALSA device will always have data ready.